Crate risc0_circuit_recursion

The recursion VM is a non-Turing-complete virtual machine (VM) optimized for algebraic constraint checking. In particular, it is well-tuned for verifying STARKs.

The recursion VM runs “recursion programs”, which define the functionality it will implement. As examples, the [lift], [join], and [resolve] programs are used by the risc0 ZkVM to compress a collection of STARK receipts for a composition into a single succinct receipt.

This is a low-level interface; users should prefer to use the risc0_zkvm crate.

Modules

control_id

layout

micro_op

Structs

CircuitImpl

This struct implements traits that are defined by code generated by the circuit definition.

Constants

CHECKED_COEFFS_PER_POLY

CIRCUIT

GLOBAL_MIX

GLOBAL_OUT

REGISTER_GROUP_ACCUM

REGISTER_GROUP_CODE

REGISTER_GROUP_CTRL

REGISTER_GROUP_DATA

Traits

Externs