risc0_circuit_recursion
Module control_id

Constants

ALLOWED_CONTROL_IDS

Control IDs allowed in the default set of recursion programs. Includes control IDs for the base set of recursion programs, and each power-of-two of the rv32im circuit, using Poseidon2.

ALLOWED_CONTROL_ROOT

Root of the Merkle tree constructed from ALLOWED_CONTROL_IDS, using Poseidon2.

BN254_IDENTITY_CONTROL_ID

Control ID for the identity recursion programs (ZKR), using Poseidon over the BN254 scalar field.

MIN_LIFT_PO2

Smallest cycle limit, as a power of two (po2), supported as a lift program.

POSEIDON2_CONTROL_IDS

Control IDs for included recursion programs (ZKRs), using Poseidon2 over BabyBear.

SHA256_CONTROL_IDS

Control IDs for included recursion programs (ZKRs), using SHA-256.